Rename truffle-external-compile to @truffle/external-compile
