+++ packages/external-compile/index.js
